refactor: rename files for TS
